+++ CHANGELOG.md
@@ -11,3 +11,5 @@
 ### Changes:
 
 ### Updates:
+
+* [PHP `7.3.5` `7.2.18` `7.1.29`](https://www.php.net/ChangeLog-7.php#7.3.5)