+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 ### Updates:
 
 * [PHP `7.3.5` `7.2.18` `7.1.29`](https://www.php.net/ChangeLog-7.php#7.3.5)
+* [Redis `5.0.5`](https://raw.githubusercontent.com/antirez/redis/5.0/00-RELEASENOTES)