+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 * [PHP `7.3.5` `7.2.18` `7.1.29`](https://www.php.net/ChangeLog-7.php#7.3.5)
 * [Redis `5.0.5`](https://raw.githubusercontent.com/antirez/redis/5.0/00-RELEASENOTES)
 * [MySQL `8.0.16`](https://dev.mysql.com/doc/relnotes/mysql/8.0/en/news-8-0-16.html)
-* [Memcached `1.5.15`](https://github.com/memcached/memcached/wiki/ReleaseNotes1515)
+* [Memcached `1.5.16`](https://github.com/memcached/memcached/wiki/ReleaseNotes1515)
 * [MariaDB `10.4.5`](https://mariadb.com/kb/en/library/mariadb-1045-release-notes/)
 * [PostgreSQL `11.3`](https://www.postgresql.org/about/news/1939/)
 * MongoDB `4.1.11`