+++ CHANGELOG.md
@@ -13,3 +13,5 @@
 ### Updates:
 
 * [PHP `7.3.6` `7.2.19` `7.1.30`](https://www.php.net/ChangeLog-7.php#7.3.6)
+* [NGINX `1.17.0`](https://nginx.org/en/CHANGES)
+* [Kubernetes `1.15.0`](https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG-1.15.md)