+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 * [NGINX `1.17.0`](https://nginx.org/en/CHANGES)
 * [Kubernetes `1.15.0`](https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG-1.15.md)
 * [PostgreSQL `11.4`](https://www.postgresql.org/about/news/1949/)
+* [MariaDB `10.4.6`](https://mariadb.com/kb/en/library/mariadb-1046-release-notes/)