+++ CHANGELOG.md
@@ -15,3 +15,5 @@
 * [PHP `7.3.6` `7.2.19` `7.1.30`](https://www.php.net/ChangeLog-7.php#7.3.6)
 * [NGINX `1.17.0`](https://nginx.org/en/CHANGES)
 * [Kubernetes `1.15.0`](https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG-1.15.md)
+* [PostgreSQL `11.4`](https://www.postgresql.org/about/news/1949/)
+* [MariaDB `10.4.6`](https://mariadb.com/kb/en/library/mariadb-1046-release-notes/)