+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 * [PHP `7.3.7` `7.2.20`](https://www.php.net/ChangeLog-7.php#7.3.7)
 * [NGINX `1.17.2`](https://nginx.org/en/CHANGES)
 * [MySQL `8.0.17`](https://dev.mysql.com/doc/relnotes/mysql/8.0/en/news-8-0-17.html)
+* [MariaDB `10.4.7`](https://mariadb.com/kb/en/library/mariadb-1047-release-notes/)