+++ CHANGELOG.md
@@ -13,6 +13,6 @@
 ### Updates:
 
 * [PHP `7.3.8` `7.2.21` `7.1.31`](https://www.php.net/ChangeLog-7.php#7.3.8)
-* [NGINX `1.17.2`](https://nginx.org/en/CHANGES)
+* [NGINX `1.17.3`](https://nginx.org/en/CHANGES)
 * [MySQL `8.0.17`](https://dev.mysql.com/doc/relnotes/mysql/8.0/en/news-8-0-17.html)
 * [MariaDB `10.4.7`](https://mariadb.com/kb/en/library/mariadb-1047-release-notes/)